@@ -28,7 +28,7 @@
     </parent>
     <groupId>ru.investbook</groupId>
     <artifactId>investbook</artifactId>
-    <version>2021.8</version>
+    <version>2021.9</version>
 
     <name>investbook</name>
     <description>Investor Accounting Book</description>
@@ -47,8 +47,8 @@
 
     <properties>
         <!-- Valid version is (0-255).(0-255).(0-65535) -->
-        <win.msi.version>21.8</win.msi.version>
-        <java.version>16</java.version>
+        <win.msi.version>21.9</win.msi.version>
+        <java.version>17</java.version>
         <!-- version 3.2.0 provided by Spring Boot 2.4.1 has bug, using version from Spring Boot 2.3.4
          TODO remove after fix -->
         <maven-resources-plugin.version>3.1.0</maven-resources-plugin.version>
@@ -83,7 +83,7 @@
         <dependency>
             <groupId>com.github.spacious-team</groupId>
             <artifactId>broker-report-parser-api</artifactId>
-            <version>2021.8</version>
+            <version>2021.9</version>
         </dependency>
         <dependency>
             <groupId>com.github.spacious-team</groupId>
