@@ -25,6 +25,7 @@ import org.spacious_team.broker.pojo.PortfolioCash;
 import org.spacious_team.broker.pojo.PortfolioProperty;
 import org.spacious_team.broker.pojo.PortfolioPropertyType;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.investbook.converter.PortfolioConverter;
 import ru.investbook.converter.PortfolioPropertyConverter;
 import ru.investbook.entity.PortfolioPropertyEntity;
@@ -59,20 +60,26 @@ public class PortfolioPropertyFormsService implements FormsService<PortfolioProp
     private final PortfolioConverter portfolioConverter;
     private final Collection<String> properties = Set.of(TOTAL_ASSETS_RUB.name(), TOTAL_ASSETS_USD.name(), CASH.name());
 
+    @Transactional(readOnly = true)
     public Optional<PortfolioPropertyModel> getById(Integer id) {
         return portfolioPropertyRepository.findById(id)
                 .map(this::toModel);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<PortfolioPropertyModel> getAll() {
-        return portfolioPropertyRepository.findByPropertyInOrderByTimestampDesc(properties)
+        return portfolioPropertyRepository
+                .findByPortfolioInAndPropertyInOrderByTimestampDesc(
+                        portfolioRepository.findByEnabledIsTrue(),
+                        properties)
                 .stream()
                 .map(this::toModel)
                 .collect(Collectors.toList());
     }
 
     @Override
+    @Transactional
     public void save(PortfolioPropertyModel m) {
         saveAndFlush(m.getPortfolio());
         PortfolioProperty.PortfolioPropertyBuilder builder = PortfolioProperty.builder()
@@ -80,8 +87,7 @@ public class PortfolioPropertyFormsService implements FormsService<PortfolioProp
                 .portfolio(m.getPortfolio())
                 .timestamp(m.getDate().atStartOfDay(zoneId).toInstant());
 
-        if (m instanceof PortfolioPropertyCashModel) {
-            PortfolioPropertyCashModel c = (PortfolioPropertyCashModel) m;
+        if (m instanceof PortfolioPropertyCashModel c) {
             PortfolioCash.PortfolioCashBuilder b = PortfolioCash.builder().section("all");
             Collection<PortfolioCash> cash = new ArrayList<>();
             cash.add(b.value(c.getCashRub()).currency("RUB").build());
@@ -94,8 +100,7 @@ public class PortfolioPropertyFormsService implements FormsService<PortfolioProp
                     .collect(Collectors.toList());
             builder.property(CASH)
                     .value(PortfolioCash.serialize(cash));
-        } else if (m instanceof PortfolioPropertyTotalAssetsModel) {
-            PortfolioPropertyTotalAssetsModel a = (PortfolioPropertyTotalAssetsModel) m;
+        } else if (m instanceof PortfolioPropertyTotalAssetsModel a) {
             builder.property(a.getTotalAssetsCurrency().toPortfolioProperty())
                     .value(a.getTotalAssets().toString());
         } else {
@@ -104,7 +109,7 @@ public class PortfolioPropertyFormsService implements FormsService<PortfolioProp
 
         PortfolioPropertyEntity entity = portfolioPropertyConverter.toEntity(builder.build());
         entity = portfolioPropertyRepository.save(entity);
-        m.setId(entity.getId());
+        m.setId(entity.getId()); // used in view
         portfolioPropertyRepository.flush();
     }
 
@@ -165,6 +170,7 @@ public class PortfolioPropertyFormsService implements FormsService<PortfolioProp
         }
     }
 
+    @Transactional
     public void delete(Integer id) {
         portfolioPropertyRepository.deleteById(id);
         portfolioPropertyRepository.flush();
