@@ -22,12 +22,14 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
+import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.Basic;
 import javax.persistence.Column;
-import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
@@ -36,12 +38,23 @@ import java.time.Instant;
 @Entity
 @Table(name = "transaction")
 @Data
-@EqualsAndHashCode(of = "pk")
+@EqualsAndHashCode(of = "id")
 @ToString //(exclude = {"transactionCashFlows"})
 public class TransactionEntity {
 
-    @EmbeddedId
-    private TransactionEntityPK pk;
+    @Id
+    @GenericGenerator(name = "UseExistingOrGenerateIdGenerator", strategy = "ru.investbook.entity.UseExistingOrGenerateIdGenerator")
+    @GeneratedValue(generator = "UseExistingOrGenerateIdGenerator")
+    @Column(name = "id")
+    private Integer id;
+
+    @Basic(optional = false)
+    @Column(name = "trade_id")
+    private String tradeId;
+
+    @Basic(optional = false)
+    @Column(name = "portfolio")
+    private String portfolio;
 
 //    @ManyToOne(fetch = FetchType.LAZY)
 //    @JoinColumn(name = "portfolio", referencedColumnName = "id")
@@ -53,11 +66,11 @@ public class TransactionEntity {
     @JsonIgnoreProperties({"hibernateLazyInitializer"})
     private SecurityEntity security;
 
-    @Basic
+    @Basic(optional = false)
     @Column(name = "timestamp")
     private Instant timestamp;
 
-    @Basic
+    @Basic(optional = false)
     @Column(name = "count")
     private int count;
 
