@@ -36,8 +36,13 @@ public class PsbBrokerReportFactory extends AbstractBrokerReportFactory {
     private final Pattern expectedFileNamePattern = Pattern.compile("^[0-9]+\\([^()]+\\)[0-9\\-]+\\.xls(x)?$");
 
     @Override
+    public boolean canCreate(String excelFileName, InputStream is) {
+        return super.canCreate(expectedFileNamePattern, excelFileName, is);
+    }
+
+    @Override
     public BrokerReport create(String excelFileName, InputStream is) {
-        BrokerReport brokerReport = create(expectedFileNamePattern, excelFileName, is, PsbBrokerReport::new);
+        BrokerReport brokerReport = create(excelFileName, is, PsbBrokerReport::new);
         if (brokerReport != null) {
             log.info("Обнаружен отчет '{}' фондового и срочного рынков Промсвязьбанк брокера", excelFileName);
         }
