@@ -36,9 +36,13 @@ public class SberCashAndSecurityBrokerReportFactory extends AbstractBrokerReport
     private final Pattern expectedFileNamePattern = Pattern.compile("^Зачисления[_-]и[_-]Списания[_-].*");
 
     @Override
+    public boolean canCreate(String excelFileName, InputStream is) {
+        return super.canCreate(expectedFileNamePattern, excelFileName, is);
+    }
+
+    @Override
     public BrokerReport create(String excelFileName, InputStream is) {
-        BrokerReport brokerReport = create(expectedFileNamePattern, excelFileName, is,
-                SberCashAndSecurityBrokerReportAdapter::new);
+        BrokerReport brokerReport = create(excelFileName, is, SberCashAndSecurityBrokerReportAdapter::new);
         if (brokerReport != null) {
             log.info("Обнаружен отчет зачислений и списаний '{}' СберБанк брокера", excelFileName);
         }
