@@ -73,12 +73,15 @@ public class InternalRateOfReturn {
      * @param quote may be null only if current security position is zero
      * @return internal rate of return if can be calculated or null otherwise
      */
-    public Double calc(Collection<String> portfolios, Security security, SecurityQuote quote, ViewFilter filter) {
+    public Double calc(
+            Collection<String> portfolios, Security security, SecurityQuote quote, Instant fromDate, Instant toDate) {
+
         try {
             if (getSecurityType(security.getId()) == DERIVATIVE) {
                 return null;
             }
-            FifoPositions positions = positionsFactory.get(portfolios, security, filter);
+            FifoPositionsFilter pf = FifoPositionsFilter.of(portfolios, fromDate, toDate);
+            FifoPositions positions = positionsFactory.get(security, pf);
             int count = positions.getCurrentOpenedPositionsCount();
             if (count != 0 && (quote == null || quote.getDirtyPriceInCurrency() == null)) {
                 return null;
@@ -111,8 +114,7 @@ public class InternalRateOfReturn {
     private String getTransactionCurrency(FifoPositions positions) {
         return positions.getTransactions()
                 .stream()
-                .map(t -> transactionCashFlowRepository
-                        .findByPkPortfolioAndPkTransactionIdAndPkType(t.getPortfolio(), t.getId(), PRICE.getId()))
+                .map(t -> transactionCashFlowRepository.findByTransactionIdAndCashFlowType(t.getId(), PRICE))
                 .flatMap(Optional::stream)
                 .map(TransactionCashFlowEntity::getCurrency)
                 .findAny()
@@ -146,7 +148,7 @@ public class InternalRateOfReturn {
     private Optional<BigDecimal> getTransactionValue(Transaction t, String toCurrency) {
         BigDecimal value = null;
         if (t.getId() != null) { // bond redemption, accounted by other way, skipping
-            value = transactionCashFlowRepository.findByPkPortfolioAndPkTransactionId(t.getPortfolio(), t.getId())
+            value = transactionCashFlowRepository.findByTransactionId(t.getId())
                     .stream()
                     .map(entity -> convertToCurrency(entity.getValue(), entity.getCurrency(), toCurrency))
                     .reduce(BigDecimal.ZERO, BigDecimal::add);
