@@ -28,6 +28,7 @@ import org.springframework.stereotype.Component;
 import ru.investbook.converter.SecurityConverter;
 import ru.investbook.entity.SecurityEntity;
 import ru.investbook.report.FifoPositionsFactory;
+import ru.investbook.report.FifoPositionsFilter;
 import ru.investbook.report.Table;
 import ru.investbook.report.TableFactory;
 import ru.investbook.report.ViewFilter;
@@ -97,14 +98,15 @@ public class DerivativesMarketTotalProfitExcelTableFactory implements TableFacto
         if (!currency.equalsIgnoreCase("RUB")) {
             return emptyList();
         }
+        ViewFilter filter = ViewFilter.get();
         Collection<String> contracts = portfolios.isEmpty() ?
                 transactionRepository.findDistinctDerivativeByTimestampBetweenOrderByTimestampDesc(
-                        ViewFilter.get().getFromDate(),
-                        ViewFilter.get().getToDate()) :
+                        filter.getFromDate(),
+                        filter.getToDate()) :
                 transactionRepository.findDistinctDerivativeByPortfolioInAndTimestampBetweenOrderByTimestampDesc(
                         portfolios,
-                        ViewFilter.get().getFromDate(),
-                        ViewFilter.get().getToDate());
+                        filter.getFromDate(),
+                        filter.getToDate());
 
         return contracts.stream()
                 .map(moexDerivativeCodeService::getContractGroup)
@@ -173,8 +175,9 @@ public class DerivativesMarketTotalProfitExcelTableFactory implements TableFacto
 
     private Deque<Transaction> getTransactions(Collection<String> portfolios, Set<Security> contracts) {
         ViewFilter filter = ViewFilter.get();
+        FifoPositionsFilter pf = FifoPositionsFilter.of(portfolios, filter.getFromDate(), filter.getToDate());
         return contracts.stream()
-                .map(contract -> positionsFactory.getTransactions(portfolios, contract.getId(), filter))
+                .map(contract -> positionsFactory.getTransactions(contract.getId(), pf))
                 .flatMap(Collection::stream)
                 .sorted(Comparator.comparing(Transaction::getTimestamp))
                 .collect(toCollection(LinkedList::new));
@@ -183,7 +186,8 @@ public class DerivativesMarketTotalProfitExcelTableFactory implements TableFacto
     private Instant getLastEventDate(Collection<String> portfolios, Collection<Security> contracts) {
         ViewFilter filter = ViewFilter.get();
         return contracts.stream()
-                .map(contract -> securityProfitService.getLastEventTimestamp(portfolios, contract, paymentEvents, filter))
+                .map(contract -> securityProfitService.getLastEventTimestamp(
+                        portfolios, contract, paymentEvents, filter.getFromDate(), filter.getToDate()))
                 .flatMap(Optional::stream)
                 .max(Comparator.naturalOrder())
                 .orElse(null);
