@@ -35,6 +35,7 @@ import ru.investbook.converter.SecurityConverter;
 import ru.investbook.converter.SecurityQuoteConverter;
 import ru.investbook.report.FifoPositions;
 import ru.investbook.report.FifoPositionsFactory;
+import ru.investbook.report.FifoPositionsFilter;
 import ru.investbook.report.ForeignExchangeRateService;
 import ru.investbook.report.InternalRateOfReturn;
 import ru.investbook.report.PositionHistory;
@@ -120,42 +121,43 @@ public class PortfolioStatusExcelTableFactory implements TableFactory {
 
     private Collection<String> getSecuritiesId(Collection<String> portfolios, String currency) {
         Collection<String> contracts = new ArrayList<>();
+        ViewFilter filter = ViewFilter.get();
         if (portfolios.isEmpty()) {
             contracts.addAll(
                     transactionRepository.findDistinctSecurityByCurrencyAndTimestampBetweenOrderByTimestampDesc(
                             currency,
-                            ViewFilter.get().getFromDate(),
-                            ViewFilter.get().getToDate()));
+                            filter.getFromDate(),
+                            filter.getToDate()));
             contracts.addAll(
                     transactionRepository.findDistinctFxCurrencyPairByCurrencyAndTimestampBetween(
                             currency,
-                            ViewFilter.get().getFromDate(),
-                            ViewFilter.get().getToDate()));
+                            filter.getFromDate(),
+                            filter.getToDate()));
             if (currency.equalsIgnoreCase("RUB")) {
                 contracts.addAll(
                         transactionRepository.findDistinctDerivativeByTimestampBetweenOrderByTimestampDesc(
-                                ViewFilter.get().getFromDate(),
-                                ViewFilter.get().getToDate()));
+                                filter.getFromDate(),
+                                filter.getToDate()));
             }
         } else {
             contracts.addAll(
                     transactionRepository.findDistinctSecurityByPortfolioInAndCurrencyAndTimestampBetweenOrderByTimestampDesc(
                             portfolios,
                             currency,
-                            ViewFilter.get().getFromDate(),
-                            ViewFilter.get().getToDate()));
+                            filter.getFromDate(),
+                            filter.getToDate()));
             contracts.addAll(
                     transactionRepository.findDistinctFxCurrencyPairByPortfolioInAndCurrencyAndTimestampBetween(
                             portfolios,
                             currency,
-                            ViewFilter.get().getFromDate(),
-                            ViewFilter.get().getToDate()));
+                            filter.getFromDate(),
+                            filter.getToDate()));
             if (currency.equalsIgnoreCase("RUB")) {
                 contracts.addAll(
                         transactionRepository.findDistinctDerivativeByPortfolioInAndTimestampBetweenOrderByTimestampDesc(
                                 portfolios,
-                                ViewFilter.get().getFromDate(),
-                                ViewFilter.get().getToDate()));
+                                filter.getFromDate(),
+                                filter.getToDate()));
             }
         }
         return contracts;
@@ -220,7 +222,8 @@ public class PortfolioStatusExcelTableFactory implements TableFactory {
         row.put(TYPE, securityType.getDescription());
         try {
             ViewFilter filter = ViewFilter.get();
-            FifoPositions positions = positionsFactory.get(portfolios, security, filter);
+            FifoPositionsFilter pf = FifoPositionsFilter.of(portfolios, filter.getFromDate(), filter.getToDate());
+            FifoPositions positions = positionsFactory.get(security, pf);
             row.put(FIRST_TRANSACTION_DATE, Optional.ofNullable(positions.getPositionHistories().peekFirst())
                     .map(PositionHistory::getInstant)
                     .orElse(null));
@@ -229,7 +232,8 @@ public class PortfolioStatusExcelTableFactory implements TableFactory {
                     .orElse(null));
             if (securityType != CURRENCY_PAIR) {
                 row.put(LAST_EVENT_DATE,
-                        securityProfitService.getLastEventTimestamp(portfolios, security, paymentEvents, filter)
+                        securityProfitService.getLastEventTimestamp(
+                                        portfolios, security, paymentEvents, filter.getFromDate(), filter.getToDate())
                                 .orElse(null));
             }
             row.put(BUY_COUNT, positions.getTransactions()
@@ -261,7 +265,7 @@ public class PortfolioStatusExcelTableFactory implements TableFactory {
                         .abs()
                         .divide(BigDecimal.valueOf(Math.max(1, Math.abs(count))), 6, RoundingMode.CEILING));
 
-                quote = securityProfitService.getSecurityQuote(security, toCurrency, filter);
+                quote = securityProfitService.getSecurityQuote(security, toCurrency, filter.getToDate());
 
                 if (quote != null) {
                     row.put(LAST_PRICE, quote.getCleanPriceInCurrency());
@@ -286,7 +290,8 @@ public class PortfolioStatusExcelTableFactory implements TableFactory {
                 row.put(TAX, securityProfitService.sumPaymentsForType(portfolios, security, CashFlowType.TAX, toCurrency).abs());
             }
             row.put(PROFIT, PROFIT_FORMULA);
-            row.put(INTERNAL_RATE_OF_RETURN, internalRateOfReturn.calc(portfolios, security, quote, filter));
+            row.put(INTERNAL_RATE_OF_RETURN, internalRateOfReturn.calc(
+                    portfolios, security, quote, filter.getFromDate(), filter.getToDate()));
             row.put(PROFIT_PROPORTION, PROFIT_PROPORTION_FORMULA);
         } catch (Exception e) {
             log.error("Ошибка при формировании агрегированных данных по бумаге {}", security, e);
