@@ -19,15 +19,18 @@
 package ru.investbook.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.transaction.annotation.Transactional;
 import ru.investbook.entity.EventCashFlowEntity;
+import ru.investbook.entity.PortfolioEntity;
 
 import java.time.Instant;
 import java.util.Collection;
 import java.util.List;
 
+@Transactional(readOnly = true)
 public interface EventCashFlowRepository extends JpaRepository<EventCashFlowEntity, Integer> {
 
-    List<EventCashFlowEntity> findByOrderByPortfolioIdAscTimestampDesc();
+    List<EventCashFlowEntity> findByPortfolioInOrderByPortfolioIdAscTimestampDesc(Collection<PortfolioEntity> portfolios);
 
     List<EventCashFlowEntity> findByPortfolioIdAndCashFlowTypeIdOrderByTimestamp(String portfolio,
                                                                                  int cashFlowType);
