@@ -24,26 +24,30 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.transaction.annotation.Transactional;
 import ru.investbook.entity.SecurityDescriptionEntity;
 
-
-
 @Transactional(readOnly = true)
 public interface SecurityDescriptionRepository extends JpaRepository<SecurityDescriptionEntity, String> {
 
     @Transactional
     default void createOrUpdateSector(String securityId, String sector) {
-        findById(securityId)
-                .ifPresentOrElse(
-                        $ -> updateSector(securityId, sector),
-                        () -> {
-                            var entity = new SecurityDescriptionEntity();
-                            entity.setSecurity(securityId);
-                            entity.setSector(sector);
-                            save(entity);
-                        });
+        if (existsById(securityId)) {
+            updateSector(securityId, sector);
+        } else {
+            createSectorIfNotExists(securityId, sector);
+        }
     }
 
     @Transactional
     @Modifying
     @Query("UPDATE SecurityDescriptionEntity SET sector = :sector WHERE security = :securityId")
-    int updateSector(String securityId, String sector);
+    void updateSector(String securityId, String sector);
+
+    @Transactional
+    default void createSectorIfNotExists(String securityId, String sector) {
+        if (!existsById(securityId)) {
+            var entity = new SecurityDescriptionEntity();
+            entity.setSecurity(securityId);
+            entity.setSector(sector);
+            save(entity);
+        }
+    }
 }
