@@ -46,7 +46,7 @@ public class HomePageController {
     @GetMapping
     public String index(Model model) {
         Set<String> portfolios = assetsAndCashService.getActivePortfolios();
-        model.addAttribute("transactionsCount", transactionRepository.countByPkPortfolioIn(portfolios));
+        model.addAttribute("transactionsCount", transactionRepository.countByPortfolioIn(portfolios));
         model.addAttribute("portfolios", portfolios);
         model.addAttribute("assets", assetsAndCashService.getAssets(portfolios));
         model.addAttribute("cashBalance", assetsAndCashService.getTotalCash(portfolios));
