@@ -23,6 +23,7 @@ import org.spacious_team.broker.pojo.CashFlowType;
 import org.spacious_team.broker.pojo.EventCashFlow;
 import org.spacious_team.broker.pojo.Portfolio;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 import ru.investbook.converter.EventCashFlowConverter;
 import ru.investbook.converter.PortfolioConverter;
@@ -45,20 +46,24 @@ public class EventCashFlowFormsService implements FormsService<EventCashFlowMode
     private final EventCashFlowConverter eventCashFlowConverter;
     private final PortfolioConverter portfolioConverter;
 
+    @Transactional(readOnly = true)
     public Optional<EventCashFlowModel> getById(Integer id) {
         return eventCashFlowRepository.findById(id)
                 .map(this::toModel);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<EventCashFlowModel> getAll() {
-        return eventCashFlowRepository.findByOrderByPortfolioIdAscTimestampDesc()
+        return eventCashFlowRepository
+                .findByPortfolioInOrderByPortfolioIdAscTimestampDesc(portfolioRepository.findByEnabledIsTrue())
                 .stream()
                 .map(this::toModel)
                 .collect(Collectors.toList());
     }
 
     @Override
+    @Transactional
     public void save(EventCashFlowModel e) {
         saveAndFlush(e.getPortfolio());
         EventCashFlowEntity entity = eventCashFlowRepository.save(
@@ -71,7 +76,7 @@ public class EventCashFlowFormsService implements FormsService<EventCashFlowMode
                         .currency(e.getValueCurrency())
                         .description(StringUtils.hasText(e.getDescription()) ? e.getDescription() : null)
                         .build()));
-        e.setId(entity.getId());
+        e.setId(entity.getId()); // used in view
         eventCashFlowRepository.flush();
     }
 
@@ -97,6 +102,7 @@ public class EventCashFlowFormsService implements FormsService<EventCashFlowMode
         return m;
     }
 
+    @Transactional
     public void delete(Integer id) {
         eventCashFlowRepository.deleteById(id);
         eventCashFlowRepository.flush();
